+++ src/main/java/com/transferwise/openbanking/client/oauth/domain/GetAccessTokenRequest.java
@@ -26,6 +26,11 @@ public class GetAccessTokenRequest {
     private final Map<String, String> requestBody = new HashMap<>();
     private final FapiHeaders requestHeaders = FapiHeaders.defaultHeaders();
 
+    public static GetAccessTokenRequest clientCredentialsRequest() {
+        return new GetAccessTokenRequest()
+            .setGrantType(CLIENT_CREDENTIALS_GRANT_TYPE);
+    }
+
     public static GetAccessTokenRequest clientCredentialsRequest(String scope) {
         return new GetAccessTokenRequest()
             .setGrantType(CLIENT_CREDENTIALS_GRANT_TYPE)